fix: UserRole ID Long타입으로 변경

+++ NarangNorang/src/main/java/com/narangnorang/user/entity/UserRole.java
@@ -1,9 +1,6 @@
 package com.narangnorang.user.entity;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
+import jakarta.persistence.*;
 import lombok.Data;
 import lombok.Getter;
 
@@ -13,6 +10,6 @@ import lombok.Getter;
 public class UserRole {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String name;
 }
